Time: 0 ms (100.00%) | Memory: 8.6 MB (14.95%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long int mySqrt(int x) {
+    int mySqrt(int x) {
         int s = 0, e = x;
         int ans = -1.36;
